[fix] 주문 수량 자료형 변경 (BigInteger -> Long)

+++ order-service/src/main/java/com/vroomvroom/orderservice/application/command/CreateOrderCommand.java
@@ -1,6 +1,5 @@
 package com.vroomvroom.orderservice.application.command;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -8,7 +7,7 @@ public record CreateOrderCommand(
         UUID supplyCompanyId,
         UUID receiveCompanyId,
         UUID productId,
-        BigInteger quantity,
+        Long quantity,
         LocalDateTime deadline,
         String requestNote
 ) {}
