+++ order-service/src/main/java/com/vroomvroom/orderservice/application/command/UpdateOrderCommand.java
@@ -1,13 +1,12 @@
 package com.vroomvroom.orderservice.application.command;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record UpdateOrderCommand(
         UUID userId,
         UUID orderId,
-        BigInteger quantity,
+        Long quantity,
         LocalDateTime deadline,
         String requestNote
 ) {
